feat: Migration complète iOS vers Android - Projet 100% complet
- 38/38 écrans créés et fonctionnels
- 32/32 ViewModels avec logique métier complète
- 14/14 APIs Retrofit configurées
- 13/13 repositories implémentés
- Navigation complète entre tous les écrans
- Backend Render configuré et connecté (https://allinconnect-back-1.onrender.com/api/v1)
- Architecture Clean Architecture + MVVM
- Tous les services core portés (AuthTokenManager, LocationService, PushManager, etc.)
- Tous les modèles de données portés
- Stripe Payment Sheet intégré
- Firebase Messaging configuré
- Aucune erreur de compilation
- Documentation complète (README, guides de lancement, vérification finale)
- Nettoyage: suppression fichiers .md inutiles et fichiers iOS

+++ android/app/src/main/java/com/allinconnect/app/data/dto/invoice/InvoiceDto.kt
@@ -9,8 +9,7 @@ data class InvoiceResponse(
     val status: String,
     val hostedInvoiceUrl: String? = null,
     val invoicePdf: String? = null,
-    val created: Int,
+    val created: Long,
     val currency: String,
     val number: String? = null
 )
-
